requirements.txt
Django==1.8.2
django-ipware==1.0.0
djangorestframework==3.1.2
requests==2.7.0
wsgiref==0.1.2
